+++ docs/CHANGELOG.md
@@ -6,6 +6,9 @@
 **Gameplay**
 - Added the ability to combine the two halves of an item in the inventory, such as the eye pieces in Tomb of Seth (TRX482)
 
+**TR2**
+- Fixed missing gun flashes in the Diving Area cutscene (OG bug) (TRX89)
+
 **TR4**
 - Added Small Scorpion control (TRX1562)
 - Added Jackal control (TRX1553)